Step 1: Parse Data Files

Step 1: Parse Data Files

Usage

threader_parse(data_path = NULL, spec_fpath = NULL, verbose = FALSE)

Arguments

data_path

Path to the directory containing the raw data you want to parse.

spec_fpath

Filepath to the .yaml-format spec file

verbose

Optional: If TRUE, prints additional debugging information.

Value

An environment where each key is an input file, and each value is the tibble generated from the data in that file

Examples

all_dfs <- threader_parse()
#> [1] "Parsing /home/runner/work/_temp/Library/threadsheetr/extdata//cp_membership_num/1924-1968_hsssd_cp_membership.csv"
#> [1] "Detected 2 headers"
#> [1] "num_headers: 2"
#> New names:
#> • `` -> `...2`
#> • `` -> `...3`
#> • `` -> `...4`
#> • `` -> `...8`
#> Warning: NAs introduced by coercion
#> [1] "Parsing /home/runner/work/_temp/Library/threadsheetr/extdata//cp_membership_num/1947-1976_statistical_abstract_latin_america.csv"
#> [1] "Detected 2 headers"
#> [1] "num_headers: 2"
#> New names:
#> • `30000` -> `30000...2`
#> • `30000` -> `30000...3`
#> • `50000` -> `50000...4`
#> • `80000` -> `80000...5`
#> • `80000` -> `80000...6`
#> • `80000` -> `80000...7`
#> • `80000` -> `80000...8`
#> • `50000` -> `50000...9`
#> • `50000` -> `50000...10`
#> • `65000` -> `65000...12`
#> • `65000` -> `65000...13`
#> • `60000` -> `60000...14`
#> • `60000` -> `60000...15`
#> • `60000` -> `60000...16`
#> • `60000` -> `60000...18`
#> • `60000` -> `60000...19`
#> Warning: NAs introduced by coercion
#> [1] "Parsing /home/runner/work/_temp/Library/threadsheetr/extdata//cp_membership_num/1963-1973_CP_Strength_State_Dept.csv"
#> [1] "Detected 0 headers"
#> [1] "num_headers: 0"
#> Warning: NAs introduced by coercion
#> [1] "Parsing /home/runner/work/_temp/Library/threadsheetr/extdata//cp_membership_num/1971-1990_Hoover_Yearbooks.csv"
#> [1] "Detected 1 headers"
#> [1] "num_headers: 1"
#> New names:
#> • `` -> `...3`
#> • `` -> `...10`
#> • `` -> `...11`
#> • `` -> `...12`
#> • `` -> `...13`
#> • `` -> `...16`
#> • `` -> `...17`
#> • `` -> `...18`
#> • `` -> `...19`
#> • `` -> `...20`
#> • `` -> `...21`
#> • `` -> `...22`
#> • `` -> `...23`
#> Warning: NAs introduced by coercion